Bug 906698 - nsUTF16ToUnicode is mis-converting on big-endian processor. r=smontagu CLOSED TREE
UltraBlame original commit: 2654e3d27d419b83294ff1ac3f49a5a9b94cfcb0

+++ intl/uconv/ucvlatin/nsUTF16ToUnicode.cpp
@@ -78,7 +78,7 @@ nsUTF16ToUnicodeBase::UTF16ConvertToUnicode(const char * aSrc,
     
     mState = STATE_NORMAL;
 #ifdef IS_BIG_ENDIAN
-    u = (mOddByte << 8) | *src++; 
+    u = (mOddByte << 8) | uint8_t(*src++); 
 #else
     u = (*src++ << 8) | mOddByte; 
 #endif
